+++ src/compile-middleware.ts
@@ -167,7 +167,9 @@ function compileHtml(
     const src = dom5.getAttribute(scriptTag, 'src');
     const isInline = !src;
 
-    if (src && src.includes('web-component-tester/browser.js')) {
+    if (src &&
+        (src.includes('web-component-tester/browser.js') ||
+         src.includes('wct-browser-legacy/browser.js'))) {
       wctScriptTag = scriptTag;
     }
 
@@ -265,9 +267,9 @@ function compileHtml(
     var moduleCount = 0;
     window.require = function(deps, factory) {
       moduleCount++;
-      originalRequire(deps, function(...args) {
+      originalRequire(deps, function() {
         if (factory) {
-          factory(...args);
+          factory.apply(undefined, arguments);
         }
         moduleCount--;
         if (moduleCount === 0) {